requirement.txt
amqp==2.5.1
apns2==0.7.1
astroid==2.2.5
authy==2.2.3
autopep8==1.4.4
billiard==3.6.1.0
CacheControl==0.12.5
cachetools==3.1.1
celery==4.3.0
certifi==2018.11.29
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
Django==2.1.5
django-cors-headers==2.5.2
django-filter==2.1.0
django-filters==0.2.1
django-push-notifications==2.0.0
django-silk==3.0.2
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
firebase-admin==3.0.0
google-api-core==1.14.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.4.0
google-cloud-storage==1.20.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
gprof2dot==2017.9.19
grpcio==1.24.1
gunicorn==19.9.0
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
httplib2==0.14.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.8
importlib-metadata==0.23
isort==4.3.15
Jinja2==2.10.1
kombu==4.6.5
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
msgpack==0.6.2
mysqlclient==1.4.1
Pillow==7.1.1
protobuf==3.10.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
py-vapid==1.7.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycodestyle==2.5.0
pycparser==2.20
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.3.1
python-dateutil==2.8.0
pytz==2018.9
pywebpush==1.10.1
requests==2.21.0
rsa==4.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
stripe==2.24.1
typed-ast==1.3.1
uritemplate==3.0.0
urllib3==1.24.1
vine==1.3.0
wrapt==1.11.1
zipp==0.6.0
